1.20

+++ src/main/java/me/aleksilassila/litematica/printer/mixin/openinv/MixinBlock.java
@@ -5,6 +5,7 @@
 
 package me.aleksilassila.litematica.printer.mixin.openinv;
 
+import me.aleksilassila.litematica.printer.LitematicaMixinMod;
 import me.aleksilassila.litematica.printer.printer.memory.MemoryDatabase;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -30,6 +31,7 @@ public abstract class MixinBlock {
             at = {@At("TAIL")}
     )
     private void chestTracker$handleBlockBreak(World world, BlockPos pos, BlockState state, PlayerEntity player, CallbackInfo ci) {
+        if(!LitematicaMixinMod.INVENTORY.getBooleanValue()) return;
         MemoryDatabase database = MemoryDatabase.getCurrent();
         if (database != null) {
             database.removePos(world.getRegistryKey().getValue(), pos);
